requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.1
djangorestframework==3.15.2
gunicorn==22.0.0
dj_database_url==0.5.0
whitenoise==6.8.2
psycopg2==2.9.10
django-cors-headers==4.6.0 
Pillow==11.0.0
python-dotenv~=1.0.1